Adding model loader
