+++ handwriting_ocr/data/data_loader.py
@@ -2,102 +2,18 @@
 # Licensed under the MIT License. See LICENSE for details.
 """Modelu for downloading datasets and loading data (DATASETS list)."""
 
-from abc import ABCMeta, abstractmethod
-import getpass
-import gzip
-from pathlib import Path
-import tarfile
-import urllib.request
 import xml.etree.ElementTree
-import zipfile
+from abc import ABCMeta, abstractmethod
+from pathlib import Path
 
 import cv2 as cv
-import gdown
 import numpy as np
-from tqdm import tqdm
-
+from handwriting_ocr.data.loader import Loader
 
 DATA_FOLDER = Path(__file__).parent.joinpath("../../data/")
 
 
-class Progressbar(tqdm):
-    """Helper class for download progressbar."""
-
-    def update_to(self, b=1, bsize=1, tsize=None):
-        if tsize is not None:
-            self.total = tsize
-        self.update(b * bsize - self.n)
-
-
-def download_url(url, output, username=None, password=None):
-    """Download file from URL to output location.
-
-    Args:
-        url (str): URL for downloading the file
-        output (Path): Path where should be downloaded file stored
-        username (str): (Optional) username for authentication
-        password (str): (Optional) username for authentication
-
-    Returns:
-        status (int): Returns status of download (200: OK, 401: Unauthorized,
-            -1: Unknown)
-    """
-
-    if "drive.google.com" in url:
-        gdown.download(url, str(output), quiet=False)
-        return 200
-
-    for _ in range(3):
-        if username or password:
-            # create a password manager
-            password_mgr = urllib.request.HTTPPasswordMgrWithDefaultRealm()
-            password_mgr.add_password(None, url, username, password)
-            handler = urllib.request.HTTPBasicAuthHandler(password_mgr)
-            # create "opener"
-            opener = urllib.request.build_opener(handler)
-            urllib.request.install_opener(opener)
-
-        with Progressbar(
-            unit="B", unit_scale=True, miniters=1, desc=url.split("/")[-1]
-        ) as t:
-            try:
-                urllib.request.urlretrieve(url, filename=output, reporthook=t.update_to)
-                return 200
-            except urllib.error.HTTPError as e:
-                if hasattr(e, "code") and e.code == 401:
-                    return 401
-                print(f"\nError occured during download:\n{e}")
-                return -1
-
-
-def file_extract(file_path, out_path):
-    """Extract archive file into given location.
-
-    Args:
-        file_path (Path): Path of archive file
-        out_path (Path): Path of folder where should be the file extracted
-    """
-    print(f"Extracting {file_path} file...")
-    if file_path.suffix == ".zip":
-
-        def open_file(x):
-            return zipfile.ZipFile(x, "r")
-
-    elif file_path.suffix in [".gz", ".tgz"]:
-
-        def open_file(x):
-            return tarfile.open(x, "r:gz")
-
-    elif file_path.suffix == ".tar":
-
-        def open_file(x):
-            return tarfile.open(x, "r")
-
-    with open_file(file_path) as data_file:
-        data_file.extractall(out_path)
-
-
-class Data(metaclass=ABCMeta):
+class Dataset(Loader):
     """Abstract class for managing data.
 
     Attributes:
@@ -108,19 +24,6 @@ class Data(metaclass=ABCMeta):
         username (str): (Optional) username for authentication during donwload
         password (str): (Optional) password for authentication during donwload
     """
-
-    require_auth = False
-    username, password = None, None
-
-    @property
-    @abstractmethod
-    def name(self):
-        ...
-
-    @property
-    @abstractmethod
-    def files(self):
-        ...
 
     @abstractmethod
     def load(self, data_path):
@@ -135,83 +38,15 @@ class Data(metaclass=ABCMeta):
         """
         ...
 
-    def clear(self, data_path):
-        """Clear all downloaded files of dataset.
-
-        Args:
-            data_path (Path): Path to data folder
-        """
-        for _, _, res, folder in self.files:
-            d = data_path / folder / self.name
-            if not d.exists():
-                shutil.rmtree(d)
-
-    def is_downloaded(self, data_path):
-        """Check if dataset is downloaded.
-
-        Args:
-            data_path (Path): Path to data folder
-
-        Returns:
-            is_downloaded (bool): True if dataset is downloaded (no folder missing)
-        """
-        for _, _, res, folder in self.files:
-            if not data_path.joinpath(folder, self.name, res).exists():
-                return False
-        return True
-
-    def download(self, data_path):
-        print(f"Collecting dataset {self.name}...")
-        downloaded = False
-        for url, f, res, folder in self.files:
-            folder = data_path / folder / self.name
-            tmp_output = folder / f
-            res_output = folder / res
-
-            if not res_output.exists():
-                tmp_output.parent.mkdir(parents=True, exist_ok=True)
-                # Try the authentication 3 times
-                for i in range(3):
-                    if self.require_auth:
-                        if not self.username:
-                            self.username = input(f"Username for {self.name} dataset: ")
-                        if not self.password:
-                            self.password = getpass.getpass(
-                                f"Password for {self.name} dataset: "
-                            )
-                    status = download_url(url, tmp_output, self.username, self.password)
-                    if status == 200:
-                        break
-
-                    if status == 401 and i < 2:
-                        print("Invalid username or password, please try again.")
-                    else:
-                        print(f"Dataset {self.name} skipped.")
-                        return
-                downloaded = True
-
-                if tmp_output.suffix in [".zip", ".gz", ".tgz", ".tar"]:
-                    file_extract(tmp_output, res_output)
-                    tmp_output.unlink()
-
-        if downloaded:
-            self.post_download(data_path)
-
-    def post_download(self, data_path):
-        """Run post-processing on downloaded data (e.g. cut lines from form images)
-
-        Args:
-            data_path (Path): Path to data folder
-        """
-
     def __str__(self):
-        return self.name
+        return f"dataset-{self.name}"
 
 
-class Breta(Data):
+class Breta(Dataset):
     """Handwriting data from Břetislav Hájek."""
 
     name = "breta"
+    # TODO: Remove data.zip archive (no longer in use)
     files = [
         (
             "https://drive.google.com/uc?id=1y6Kkcfk4DkEacdy34HJtwjPVa1ZhyBgg",
@@ -235,7 +70,7 @@ class Breta(Data):
         return sorted((p, p.name.split("_")[0]) for p in folder.glob("**/*.png"))
 
 
-class CVL(Data):
+class CVL(Dataset):
     """CVL Database
     More info at: https://zenodo.org/record/1492267#.Xob4lPGxXeR
     """
@@ -284,7 +119,7 @@ class CVL(Data):
         )
 
 
-class IAM(Data):
+class IAM(Dataset):
     """IAM Handwriting Database
     More info at: http://www.fki.inf.unibe.ch/databases/iam-handwriting-database
     """
@@ -324,7 +159,7 @@ class IAM(Data):
         return sorted((p, l_dic[p.with_suffix("").name]) for p in ln_f.glob("**/*.png"))
 
 
-class ORAND(Data):
+class ORAND(Dataset):
     """ORAND CAR 2014 dataset
     More info at: https://www.orand.cl/icfhr2014-hdsr/#datasets
     """
@@ -354,7 +189,7 @@ class ORAND(Data):
         return sorted(lines)
 
 
-class Camb(Data):
+class Camb(Dataset):
     """Cambridge Handwriting Database
     More info at: ftp://svr-ftp.eng.cam.ac.uk/pub/data/handwriting_databases.README
     """
@@ -406,7 +241,7 @@ class Camb(Data):
         return sorted((p, p.name.split("_")[0]) for p in folder.glob("**/*.png"))
 
 
-class NIST(Data):
+class NIST(Dataset):
     """NIST SD 19 - character dataset
     More info at: https://www.nist.gov/srd/nist-special-database-19
     """
